Moved around textured model loading and got item/doorknob texturing working.
