+++ MagicDoorknob/src/main/java/com/tomboshoven/minecraft/magicdoorknob/items/ItemMagicDoorknob.java
@@ -5,14 +5,18 @@ import com.tomboshoven.minecraft.magicdoorknob.blocks.BlockMagicDoorway;
 import com.tomboshoven.minecraft.magicdoorknob.blocks.Blocks;
 import com.tomboshoven.minecraft.magicdoorknob.blocks.tileentities.TileEntityMagicDoor;
 import com.tomboshoven.minecraft.magicdoorknob.blocks.tileentities.TileEntityMagicDoorway;
+import com.tomboshoven.minecraft.magicdoorknob.modelloaders.textured.IItemStackTextureMapperProvider;
+import com.tomboshoven.minecraft.magicdoorknob.modelloaders.textured.ITextureMapper;
 import mcp.MethodsReturnNonnullByDefault;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumActionResult;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.EnumHand;
+import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.EnumSkyBlock;
 import net.minecraft.world.IBlockAccess;
@@ -22,7 +26,7 @@ import javax.annotation.ParametersAreNonnullByDefault;
 
 @ParametersAreNonnullByDefault
 @MethodsReturnNonnullByDefault
-public class ItemMagicDoorknob extends Item {
+public class ItemMagicDoorknob extends Item implements IItemStackTextureMapperProvider {
     @Override
     public EnumActionResult onItemUse(EntityPlayer player, World worldIn, BlockPos pos, EnumHand hand, EnumFacing facing, float hitX, float hitY, float hitZ) {
         if (!worldIn.isRemote) {
@@ -122,5 +126,14 @@ public class ItemMagicDoorknob extends Item {
         }
         // TODO: Whitelist, Evaluate hardness
         return true;
+    }
+
+    @Override
+    public ITextureMapper getTextureMapper(ItemStack stack) {
+        return (spriteToMap, blockState) -> {
+            String name = spriteToMap.getIconName();
+
+            return new ResourceLocation("minecraft:blocks/redstone_block");
+        };
     }
 }
