+++ MagicDoorknob/src/main/resources/assets/magic_doorknob/blockstates/textured/magic_doorknob.json
@@ -1,0 +1,7 @@
+{
+  "variants": {
+    "normal": {
+      "model": "magic_doorknob:magic_doorknob2"
+    }
+  }
+}
