+++ MagicDoorknob/src/main/resources/assets/magic_doorknob/models/block/magic_doorknob.json
